🚚 move : productDetail 패키지 정리 및 임포트 수정
