+++ src/main/java/com/example/evostyle/domain/order/entity/OrderItem.java
@@ -1,6 +1,6 @@
 package com.example.evostyle.domain.order.entity;
 
-import com.example.evostyle.domain.product.productdetail.entity.ProductDetail;
+import com.example.evostyle.domain.product.entity.ProductDetail;
 import jakarta.persistence.*;
 import lombok.AccessLevel;
 import lombok.Getter;
