+++ src/main/java/com/example/evostyle/domain/product/repository/ProductDetailRepository.java
@@ -1,6 +1,6 @@
 package com.example.evostyle.domain.product.repository;
 
-import com.example.evostyle.domain.product.productdetail.entity.ProductDetail;
+import com.example.evostyle.domain.product.entity.ProductDetail;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ProductDetailRepository extends JpaRepository<ProductDetail,Long> {
